// src/app/modules/Service/service.interface.ts
export interface TService {
  title: string;
  slug: string;
  metaTitle: string;
  metaDescription: string;
  price: number;
  photo: string;
  currency: string;
  database1: string;
  database2: string;
  database3?: string;
  database4?: string;
  database5?: string;
  description: string;
  availability: string;
  backEndTech1: string;
  backEndTech2: string;
  backEndTech3: string;
  backEndTech4?: string;
  backEndTech5?: string;
  backEndTech6?: string;
  turnAroundTime: string;
  frontendTech1: string;
  frontendTech2: string;
  frontendTech3: string;
  frontendTech4: string;
  frontendTech5?: string;
  frontendTech6?: string;
  websiteFeature1: string;
  websiteFeature2: string;
  websiteFeature3: string;
  websiteFeature4: string;
  websiteFeature5: string;
  websiteFeature6: string;
  websiteFeature7: string;
  websiteFeature8: string;
  websiteFeature9: string;
  websiteFeature10: string;
  websiteFeature11?: string;
  websiteFeature12?: string;
  websiteFeature13?: string;
  websiteFeature14?: string;
  websiteFeature15?: string;
  securityFeatures1: string;
  securityFeatures2: string;
  securityFeatures3: string;
  securityFeatures4: string;
  securityFeatures5: string;
  securityFeatures6: string;
  securityFeatures7: string;
  securityFeatures8: string;
}